% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colar.R
\name{colar}
\alias{colar}
\title{Cola dois textos}
\usage{
colar(x, y)
}
\arguments{
\item{x}{primeiro argumento}

\item{y}{segundo arumento}
}
\value{
a cola de dois textos
}
\description{
Recebe um texto x e outro y e retorna a cola de x e y
}
