% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{soma}
\alias{soma}
\title{Soma dois números}
\usage{
soma(x, y)
}
\arguments{
\item{x}{primeiro argumento}

\item{y}{segundo arumento}
}
\value{
a soma de dois números
}
\description{
Recebe um número x e outro y e retorna a soma de x e y
}
